+++ src/const.ts
@@ -1,4 +1,4 @@
 export const Setting = {
-    FavoritesCount: 4,
-    PlacesFound: 311
+  FavoritesCount: 4,
+  PlacesFound: 311
 };
