--- src/test.ts
@@ -1,7 +1,0 @@
-type Point = {
-  x: number;
-  y: number;
-}
-
-type PointPL = keyof Point;
-let myPL: PointPL = 'y'
